third_party/WebKit/LayoutTests/media/controls-volume-slider.html

<!DOCTYPE html>
<title>media controls volume slider</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="media-controls.js"></script>
<audio controls></audio>
<script>
async_test(function(t) {
    // This test is not valid for modern media controls.
    if (internals.runtimeFlags.modernMediaControlsEnabled)
      t.done();

    var audio = document.querySelector("audio");
    audio.src = "content/test.oga";
    assert_equals(audio.volume, 1);

    audio.onloadedmetadata = t.step_func(function() {
        // click the middle of the volume slider
        var coords = mediaControlsButtonCoordinates(audio, "volume-slider");
        eventSender.mouseMoveTo(coords[0], coords[1]);
        eventSender.mouseDown();
        eventSender.mouseMoveTo(coords[0]-1, coords[1]);
        eventSender.mouseMoveTo(coords[0], coords[1]);
        eventSender.mouseUp();

        assert_greater_than(audio.volume, 0.4);
        assert_less_than(audio.volume, 0.6);
        audio.onvolumechange = t.step_func_done();
    });
});
</script>
